Time: 156 ms (74.18%) | Memory: 107.2 MB (60.42%) - LeetSync

+++ 3351-maximize-happiness-of-selected-children/maximize-happiness-of-selected-children.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long maximumHappinessSum(vector<int>& h, int k) {
         sort(h.begin(),h.end());
-        long sum=0;
+        long long sum=0;
         int min=0;
         for(int i=h.size()-1;i>=0 && k>0;i--){
             k--;
